Create Password Checker
The user will input a password and it is checked to see if it is hacked using a method known as k-anonymity. Rather than checking the password in plain text (which is dangerous, as if someone can get their hands on the data, the password is exposed), the password is converted into a hash code. However, this hash code can still be deciphered, so, using k-anonymity, only the first 5 digits of the hash are grabbed and viewed (making it almost impossible to hack).